rename model.py to get_date_matrix.py; fix test cases
