+++ tests/src/app_test.py
@@ -26,7 +26,7 @@ class TestFlaskApp(unittest.TestCase):
     def test_date_page_should_return_expected_message_for_valid_date(self, stub_get_date_matrix):
         dummy_date = "2022-02-27"
         expected_date_matrix = b"[[30, 31, 1, 2, 3, 4, 5], [6, 7, 8, 9, 10, 11, 12], [13, 14, 15, 16, 17, 18, 19], [20, 21, 22, 23, 24, 25, 26], [27, 28, 1, 2, 3, 4, 5], [6, 7, 8, 9, 10, 11, 12]]"
-        stub_get_date_matrix.return_value = json.loads(expected_date_matrix.decode())
+        stub_get_date_matrix.get_date_matrix.return_value = json.loads(expected_date_matrix.decode())
         expected_response = FakeResponse(data=expected_date_matrix, status_code=200)
 
         with self._app.test_client() as test_client:
@@ -39,7 +39,7 @@ class TestFlaskApp(unittest.TestCase):
     @patch("src.app.get_date_matrix")
     def test_date_page_should_return_400_for_invalid_date(self, stub_get_date_matrix):
         dummy_date = "2022-02-27"
-        stub_get_date_matrix.side_effect = InvalidDateFormat("unittest-invalid-date")
+        stub_get_date_matrix.get_date_matrix.side_effect = InvalidDateFormat("unittest-invalid-date")
         expected_response = FakeResponse(data="unittest-invalid-date", status_code=400)
 
         with self._app.test_client() as test_client:
@@ -52,7 +52,7 @@ class TestFlaskApp(unittest.TestCase):
     @patch("src.app.get_date_matrix")
     def test_date_page_should_return_500_for_unexpected_server_side_error(self, stub_get_date_matrix):
         dummy_date = "2022-02-27"
-        stub_get_date_matrix.side_effect = Exception("unittest-server-side-exception")
+        stub_get_date_matrix.get_date_matrix.side_effect = Exception("unittest-server-side-exception")
         expected_response = FakeResponse(data="Server side issue", status_code=500)
 
         with self._app.test_client() as test_client:
